requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1
django-admin-rangefilter==0.13.2
django-ckeditor-5==0.1.8
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
transliterate==1.10.2
urllib3==2.2.3
